+++ src/test/java/org/springframework/test/ioc/bean/Car.java
@@ -1,0 +1,24 @@
+package org.springframework.test.ioc.bean;
+
+/**
+ * @author derekyi
+ */
+public class Car {
+
+    private String brand;
+
+    public String getBrand() {
+        return brand;
+    }
+
+    public void setBrand(String brand) {
+        this.brand = brand;
+    }
+
+    @Override
+    public String toString() {
+        return "Car{" +
+            "brand='" + brand + '\'' +
+            '}';
+    }
+}