+++ src/test/java/org/springframework/test/ioc/service/HelloService.java
@@ -1,5 +1,8 @@
 package org.springframework.test.ioc.service;
 
+/**
+ * @author derekyi
+ */
 public class HelloService {
 
     public String sayHello() {