+++ changelog.md
@@ -122,3 +122,72 @@ public class PopulateBeanWithPropertyValuesTest {
 
 }
 ```
+
+## 为bean注入bean
+> 代码分支：populate-bean-with-bean
+
+增加BeanReference类，包装一个bean对另一个的bean的引用。实例化BeanA后，如果ProperValue#value为BeanReference类型，引用BeanB，则先实例化BeanB，再为BeanA设置属性。
+暂不支持循环依赖。
+
+```
+protected void applyPropertyValues(String beanName, Object bean, BeanDefinition beanDefinition) {
+    try {
+        for (PropertyValue propertyValue : beanDefinition.getPropertyValues().getPropertyValueList()) {
+            String name = propertyValue.getName();
+            Object value = propertyValue.getValue();
+
+            if (Objects.nonNull(value) && value instanceof BeanReference) {
+                // beanA依赖beanB，先实例化beanB
+                BeanReference beanReference = (BeanReference) value;
+                value = getBean(beanReference.getBeanName());
+            }
+
+            // 通过反射设置属性值
+            BeanUtil.setFieldValue(bean, name, value);
+        }
+    } catch (Exception e) {
+        throw new BeansException("Error setting property values for bean: " + beanName, e);
+    }
+}
+```
+
+
+测试：
+```
+public class PopulateBeanWithPropertyValuesTest {
+
+    /**
+     * 为bean注入bean
+     */
+    @Test
+    public void testPopulateBeanWithBean() {
+        DefaultListableBeanFactory beanFactory = new DefaultListableBeanFactory();
+
+        // 注册Car实例
+        PropertyValues propertyValuesForCar = new PropertyValues();
+        propertyValuesForCar.addPropertyValue(new PropertyValue("brand", "tesla"));
+        BeanDefinition carBeanDefinition = new BeanDefinition(Car.class, propertyValuesForCar);
+        beanFactory.registerBeanDefinition("car", carBeanDefinition);
+
+
+        // 注册Person实例
+        PropertyValues propertyValuesForPerson = new PropertyValues();
+        propertyValuesForPerson.addPropertyValue(new PropertyValue("name", "昂热"));
+        propertyValuesForPerson.addPropertyValue(new PropertyValue("age", 100));
+        // Person实例依赖Car实例
+        propertyValuesForPerson.addPropertyValue(new PropertyValue("car", new BeanReference("car")));
+        beanFactory.registerBeanDefinition("person", new BeanDefinition(Person.class, propertyValuesForPerson));
+
+        Person person = (Person) beanFactory.getBean("person");
+        System.out.println(person);
+
+        Assert.assertEquals("昂热", person.getName());
+        Assert.assertEquals(100, person.getAge());
+        Assert.assertNotNull(person.getCar());
+        Assert.assertEquals("tesla", person.getCar().getBrand());
+
+    }
+
+}
+
+```